changed request code & added check for it in onActivityResult
Using request code as 1 is pretty common and there will be conflicts, so I chose something a bit more random. Furthermore, there needs to be a check for the request code in onActivityResult, otherwise it will execute every time any (unrelated) activity in the app is finished with a result.